Fix socket leak, throttle and retry connect() attempts.
A close() was added to avoid a leak of socket handles.

In order to deal with remote server balking at a flood of connect()
attempts, two features were added. First, a limit on the number
of connections was added.  Second, connect() attempts are retried
periodically.

TODO: The limit on the number of connections should be part of the
user interface. Some diagnostic output should be eliminated. Program
tcpcat should be fixed.